prediction_app/api/requirements.txt
# API 服务依赖
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
email-validator>=2.0.0

# 模型加载和推理
joblib>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
lightgbm>=4.0.0
holidays>=0.40

# 数据库
psycopg2-binary>=2.9.0

# 日志
python-json-logger>=2.0.7
